Create a new endpoint for fees
The free mode works with a monthly limit of fees. To know this
limit it is necessary to make a GET call. Until now, the system
only allowed calls of type POST. In order to have a better
reusability, an abstract parent class has been generated for
the connection against the client.